backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
httpx==0.25.2
